requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
commonmark==0.9.1
defusedxml==0.7.1
Django==4.0.6
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
httpie==3.2.1
idna==3.3
Markdown==3.3.7
multidict==6.0.2
Pillow==9.2.0
Pygments==2.12.0
PyJWT==2.4.0
PySocks==1.7.1
pytz==2022.1
requests==2.28.1
requests-toolbelt==0.9.1
rich==12.5.1
sqlparse==0.4.2
urllib3==1.26.10
